requirements.txt
pymupdf
pytesseract
langchain
openai
numpy
pdf2image
Pillow
python-dotenv
